win_build: msvc-specific build changes + find_package
